Refactoring
- new reducer
- use reducer helper
- fixed base selectors
