+++ src/components/Uploads/UploadsConstants.js
@@ -1,0 +1,1 @@
+export const KEY = 'uploads';
