+++ src/components/Uploads/UploadsReducerOld.js
@@ -1,53 +1,3 @@
-import { handleActions } from 'redux-actions';
-import dotProp from 'dot-prop-immutable';
-import * as Actions from './UploadsActions';
-
-/**
- * Use it for debug when styling
- */
-const debug = false;
-
-const debugDefaultState = {
-  uploads: [
-    {
-      file: {
-        name: 'File uploaded',
-      },
-      status: 'done',
-      result: {
-        id: 'bla bla',
-      },
-    },
-    {
-      file: {
-        name: 'File uploading',
-      },
-      status: 'uploading',
-      progress: 0.47,
-    },
-    {
-      file: {
-        name: 'File error',
-      },
-      status: 'error',
-      error: {
-        message: 'Something went wrong',
-      },
-    },
-    {
-      file: {
-        name: 'File initializing',
-      },
-    },
-  ],
-};
-
-const normalDefaultState = {
-  uploads: [],
-};
-
-const defaultState = debug ? debugDefaultState : normalDefaultState;
-
 function updateFile(updates) {
   return (state, { payload: { file, uploadFolderId, ...restPayload } }) =>
     dotProp.set(state, 'uploads', uploads =>
