+++ src/reducers.js
@@ -4,12 +4,12 @@ import { appReducer } from 'components/App/AppReducer';
 import { notesReducer } from 'components/NotesPage/NotesPageReducer';
 import { notesMindMapReducer } from 'components/NotesMindMap/NotesMindMapReducer';
 import { searchPanelReducer } from 'components/SearchPanel/SearchPanelReducer';
-import { attachmentsReducer } from 'components/Uploads/AttachmentsReducer';
+import { uploadsReducer } from 'components/Uploads/UploadsReducer';
 
 export const reducers = combineReducers({
   app: appReducer,
   notes: notesReducer,
   notesMindMap: notesMindMapReducer,
   searchPanel: searchPanelReducer,
-  attachments: attachmentsReducer,
+  [uploadsReducer.KEY]: uploadsReducer,
 });
